made it so that user dependent test data will not be inserted if there are no users

+++ src/main/java/assessment/prepop/bootstrapping/BootstrapData.java
@@ -67,6 +67,11 @@ public class BootstrapData{
     @PostConstruct
     public void dataInsertion(){
 
+        if(getUserList().isEmpty()){
+            logger.error("No users found, testing data not inserted.");
+            return;
+        }
+
         if (properties.getKudos()){
             kudoRepository.deleteAll();
             List<Kudo> kudoList = getTestKudos();
